Finished module coding -Added new file for char -Chicken3.py is now the main visual file -Did some simple TODOs -Added a file for future calculations that will go cross file -Added a file for exploring as it will be large
