+++ exploration.py
@@ -1,0 +1,5 @@
+class Explore: #TODO: Make a window with Locations, 
+    def __init__(self,mWindow,xRes,yRes):
+        # alreadyRunning = True #Placeholder to TODO: prevent multiple Explore windows being open
+        # if(not alreadyRunning):
+        #     print("oof")
